Moved the solver to the src code and made analysis more readable
